Move test cases.
